feat(home): marketplace product of hosting alert
