+++ server/home/home-alert/src/main/java/io/holoinsight/server/home/alert/model/emuns/FunctionEnum.java
@@ -32,4 +32,8 @@ public enum FunctionEnum {
   public String getType() {
     return type;
   }
+
+  public String getDesc() {
+    return desc;
+  }
 }
