+++ server/home/home-dal/src/main/java/io/holoinsight/server/home/dal/model/MarketplaceProduct.java
@@ -34,6 +34,8 @@ public class MarketplaceProduct {
 
   private String type;
 
+  private String version;
+
   public boolean status;
 
   public String creator;
