+++ server/home/home-dal/src/main/java/io/holoinsight/server/home/dal/model/dto/MarketplaceProductDTO.java
@@ -32,6 +32,8 @@ public class MarketplaceProductDTO {
 
   public String type;
 
+  public String version;
+
   public String creator;
 
   public String modifier;
