+++ server/home/home-web/src/main/java/io/holoinsight/server/home/web/controller/MarketplaceProductFacadeImpl.java
@@ -18,6 +18,7 @@ import io.holoinsight.server.home.facade.page.MonitorPageRequest;
 import io.holoinsight.server.home.facade.page.MonitorPageResult;
 import io.holoinsight.server.home.web.common.ManageCallback;
 import io.holoinsight.server.home.web.common.ParaCheckUtil;
+import io.holoinsight.server.home.web.common.TokenUrls;
 import io.holoinsight.server.home.web.interceptor.MonitorScopeAuth;
 import io.holoinsight.server.common.J;
 import io.holoinsight.server.common.JsonResult;
@@ -42,6 +43,7 @@ import java.util.stream.Collectors;
  */
 @RestController
 @RequestMapping("/webapi/marketplace/product")
+@TokenUrls("/webapi/marketplace/product/create")
 public class MarketplaceProductFacadeImpl extends BaseFacade {
 
   @Autowired
